pyproject.toml
[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
name = "root"
version = "0.0.0"
[tool.poetry.dependencies]
aiohttp = "*"
"discord.py" = "*"
dnspython = "==1.16.0"
jinja2 = "*"
motor = "*"
pillow = "*"
psutil = "*"
python = "^3.7"
